Beginners Guide to Contributing
Since our backend is going to be programmed in Python, we’ve chosen to
use a Pyramid Program as our framework. We chose to use Tox, because it
makes managing
the virtual environment easier. We’re using a virtual environment to
keep the project dependencies from polluting the rest of our system. We
chose to use Juju, because it makes managing and scaling easier. We are
using Virtual Box for Juju so we can deploy locally for testing.
Instead of having to deploy to the public cloud, we can deploy to a
mini virtual cloud. (LXD)

These are all of the steps I took from beginning to end.